backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
passlib[bcrypt]
pyjwt
python-multipart
